custom slider positioning fixed
custom slider positioning fixed

+++ WaifuBrew/src/main/java/start/PaneFrame/SaveLoadPane.java
@@ -191,7 +191,7 @@ public class SaveLoadPane extends JPanel {
 
     private void initInteractiveUI() {
         // initializing a slider that slides the save/load panels & add mouse listeners
-        panelScrubThrough = new CustomSlider((windowSize.x / 2), (windowSize.y / 15) * 14, scrubPage, Origin.MIDDLE_BOTTOM, "");
+        panelScrubThrough = new CustomSlider((windowSize.x / 2), (int)((windowSize.y / 15.0) * 14), scrubPage, Origin.MIDDLE_BOTTOM, "");
         addMouseListener(panelScrubThrough.retrieveMouseHandler());
         addMouseMotionListener(panelScrubThrough.retrieveMouseHandler());
     }
